// apps/media/src/infrastructure/types/auth.types.ts
export type UserDetails = {
	email: string
	username: string
}

export interface ICreateAvatarPath {
	userId: string
	avatarId: string
	originalname: string
}

export interface ICreateAvatar {
	id: string
	userId: string
	uploadPath: string
	contentType: string
	height: number
	width: number
	size: number
}